Actually fixed login issue and added newitem page
The login issue was due to a simple misnamed variable, credentials
instead of access_token in the /disconnect route. I was looking
for credentials where there was none.
I also created the newItem page and relevant routes for GET and POST
requests. I managed to join the category table to the items table
in order to show the category name in the items section instead of
just the id.

TODO: EDIT and DELETE routes for category and items